feat(backend): add product schema.

+++ backend/app/common/validators.py
@@ -1,4 +1,4 @@
-def validate_price(value: float) -> float:
+def validate_price(value: int) -> int:
     if value < 0:
         raise ValueError("Price cannot be negative")
     return value
